refactor image loading into view
